+++ Changelog.md
@@ -1,5 +1,28 @@
 # Changelog
 
+2.0.0-pre.12
+- Fixed UnityEditor namespace error during player build
+
+2.0.0-pre.11
+- Added editor toolbar : switch mode, speed gear, presentation mode and open settings window
+- Cleaned up settings window
+- Added option to only navigate when Unity has focus, using PR from Jay Ott
+- Added option to toggle led when unity gains/loses focus
+
+2.0.0-pre.10
+- Merged improvement by AJ Campbell which implements framerate independent navigation
+- Added a fix for drift + recalibration button
+
+2.0.0-pre.9
+- Merged fix by cpetry to prevent exception when quitting without SpaceNavigator connected
+
+2.0.0-pre.8
+- Merged PR by Stefan Beer which fixes drifting (issue #29, #51)
+
+2.0.0-beta.7
+- Merged PR by krisrok which enables toggling the SpaceNavigator's LED
+- Merged PR by Jiapeng Chi which fixes rotation offsets between editor and play mode
+
 2.0.0-beta.6
 - Added separate layout for SpaceNavigator Wireless
 - Merged PR by Felix Herbst to match trailing spaces in manufacturer name, fixes Space Pilot Pro and possibly others
@@ -23,7 +46,7 @@
 2.0.0-beta.1
 - Changed git package url in readme to /develop branch for beta release
 
-2.0.0-alpha.2 to 8  
+2.0.0-alpha.2 to 8
 - Package manager + git experiments
 - Prepare for beta publishing
 
@@ -46,7 +69,7 @@ Fixes
 Features
 - Added option to navigate in the editor at runtime.
 - Separate axis lock settings for navigation and manipulation modes.
-- Autosave settings every 30 seconds.
+- Autosave settings every 30 seconds.  
 Fixes
 - Fixed horizon lock, coord sys and snap setting not being saved.
 - Refactored driver settings.